Optimize move generation (#9)
* Store masks, magics, shifts and attacks in a single lookup table

* Speed up king move generation by removing make/unmake move

* Refactor generateCastlingMoves

* Refactor Bitwise, PinCalculator and RayCalculator

* Use int for Move instead of short - gain from decreased memory does not outweigh cost of constant casting to int

* Run perft tests

* Add comments

+++ src/main/java/com/kelseyde/calvin/utils/notation/Notation.java
@@ -24,7 +24,7 @@ public class Notation {
         return new Move(fromNotation(startSquare), fromNotation(endSquare));
     }
 
-    public static Move fromNotation(String startSquare, String endSquare, short flag) {
+    public static Move fromNotation(String startSquare, String endSquare, int flag) {
         return new Move(fromNotation(startSquare), fromNotation(endSquare), flag);
     }
 
@@ -36,7 +36,7 @@ public class Notation {
         int startSquare = fromNotation(notation.substring(0, 2));
         int endSquare = fromNotation(notation.substring(2, 4));
 
-        short flag = Move.NO_FLAG;
+        int flag = Move.NO_FLAG;
         if (notation.length() == 5) {
             String pieceCode = notation.substring(4, 5);
             Piece promotionPieceType = PIECE_CODE_INDEX.entrySet().stream()
